==> R/1_choose_example_stations.Rmd <==
---
title: "Season Length & Data Gaps"
output: html_document
editor_options: 
  chunk_output_type: console
---
  
```{r, echo = FALSE, message=FALSE}

library(data.table)    # export example data
library(dplyr)         # data manipulation 
library(DT)            # interactive tables
library(ggplot2)       # figures
library(here)          # relative files paths
library(leaflet)       # interactive maps
library(plotly)        # interactive plots
library(purrr)         # read in data
library(RColorBrewer)  # colour scale for season length plot
library(strings)       # convert_depth_to_ordered_factor() function
library(tgc)           # season functions

theme_set(theme_light())

dt_options <- list(
      dom = 'ft',
      paging = FALSE,
      searching = TRUE,
      scrollY = "500px",
      pageLength = 500,
      columnDefs = list(list(className = 'dt-center', targets = "_all"))
)

```

```{r, echo=FALSE, message=FALSE}

# import data
dat_raw <- list.files(here("data"), pattern = "rds", full.names = TRUE) %>%
  unlist() %>%
  map_dfr(readRDS) %>%
  filter(VARIABLE == "Temperature") %>% 
  select(-WATERBODY, -LEASE, -VARIABLE, -SENSOR, -DEPLOYMENT_PERIOD, 
         -UNITS, -MOORING) %>% 
  filter(STATION != "Piper Lake", STATION != "Hourglass Lake") 

```

Explore temperature data to select example stations for the TGC paper

  - Part 1: Growing Seasons
  - Part 2: Data Gaps
  - Part 3: Growing Season Length
  - Part 4: Visualize Data Series 
  - Part 5: Final Selection

# Station Map

```{r, echo=FALSE}

dat_raw %>% 
  distinct(STATION, .keep_all = TRUE) %>% 
  leaflet() %>%
  addProviderTiles(providers$CartoDB.Positron)%>%
  addMarkers(~LONGITUDE, ~LATITUDE, label = ~STATION)

```

  - `r length(unique(dat_raw$STATION))` STATIONs have temperature data.
  - `r nrow(distinct(dat_raw, STATION, DEPTH))` STATION + DEPTH combinations have temperature data.

# Part 1: Growing Seasons {.tabset}

Filter `dat_raw` for STATION + DEPTH data series with full season of data using the `filter_in_growing_seasons()` function, with `full_season = TRUE`. This will filter out data that is not part of a growing season, and assign each remaining observation a `SEASON` label.

```{r, message=FALSE}

dat_seasons <-  dat_raw %>%
  filter_in_growing_seasons(
    full_season = TRUE,
    trend_threshold = 4,
    superchill_threshold = -0.7,
    max_season = 540
  ) 

``` 

  - `r length(unique(dat_seasons$STATION))` STATIONs have data for at least 1 full season
  - `r nrow(distinct(dat_seasons, STATION, DEPTH))` STATION + DEPTH combinations have data for at least 1 full season


# Part 2: Data Gaps 

Check `dat_seasons` for data gaps for each STATION, DEPTH, and SEASON using the `check_for_data_gaps()` function.

The function returns a table with all data gaps greater than 2 hours & prints out the number of data gaps greater than 6 hours.

Table 1 sums the gap duration for each STATION, DEPTH, and SEASON.
 
```{r}

# Table 1
gap_table <- dat_seasons %>% 
  check_for_data_gaps(STATION, SEASON, gap_length = 2, gap_warning = 6) %>% 
  group_by(STATION, SEASON, DEPTH) %>%
  summarize(TOTAL_GAP_DAYS = sum(GAP_LENGTH_DAYS)) %>%
  ungroup()

```

## Visualize Data Gaps {.tabset}

### Figure 1

```{r, echo=FALSE, fig.height=5}

cols <- get_colour_palette(gap_table)

p <- gap_table %>% 
  convert_depth_to_ordered_factor() %>% 
  arrange(desc(TOTAL_GAP_DAYS)) %>% 
  mutate(INDEX = 1:n()) %>% 
  ggplot(
    aes(INDEX, TOTAL_GAP_DAYS, col = DEPTH,
        text = paste(
          "Station: ", STATION, "\n",
          "Depth: ", DEPTH, "\n",
          "Season: ", SEASON, "\n",
          "Gap Days: ", TOTAL_GAP_DAYS))
  ) + 
  geom_point() +
  scale_color_manual(name = "Depth (m)", values = cols) 

ggplotly(p, tooltip = "text")

```


### Figure 2

```{r, echo = FALSE, message=FALSE, fig.height=6}

gap_table_filt <- gap_table %>% 
  filter(TOTAL_GAP_DAYS > 0) %>% 
  convert_depth_to_ordered_factor() 

cols <- get_colour_palette(gap_table_filt)

p <-  gap_table_filt %>% 
  ggplot(
    aes(TOTAL_GAP_DAYS, STATION, fill = DEPTH,
        text = paste(
          "Station: ", STATION, "\n",
          "Depth: ", DEPTH, "\n",
          "Season: ", SEASON, "\n",
          "Gap Days: ", TOTAL_GAP_DAYS))
  ) +
  geom_col(
    position = position_dodge2(preserve = "single", reverse = TRUE, padding = 0)
  ) +
  scale_fill_manual(name = "Depth (m)", values = cols) + 
  scale_y_discrete(name = "", limits = rev) +
  theme(panel.grid.major.y =  element_blank()) 

ggplotly(p, tooltip = "text") %>% 
  config(
    modeBarButtonsToRemove = list(
      "toggleSpikelines", "autoScale2d", "select2d", "lasso2d",
      "hoverCompareCartesian", "zoom2d"
    ),
    displaylogo = FALSE,
    toImageButtonOptions = list(
      format = "png",
      filename = "Gap_days_station_depth",
      width = 1100,
      height = 1100
    )) %>%
  layout(
    margin = list(b = 90, l = 90) # to fully display the x and y axis labels
  )

```

### Table 1

```{r, echo=FALSE}

gap_table %>% 
  arrange(desc(TOTAL_GAP_DAYS)) %>% 
  datatable(rownames = TRUE, options = dt_options)

```


Filter out STATION + DEPTH + SEASON with data gap greater than 2 days.

```{r}

dat_seasons <- dat_seasons %>% 
  left_join(gap_table, by = c("STATION", "DEPTH", "SEASON")) %>% 
  filter(TOTAL_GAP_DAYS < 2)

```

  - `r length(unique(dat_seasons$STATION))` STATIONs have data for at least 1 full season of data with no major data gaps.
  - `r nrow(distinct(dat_seasons, STATION, DEPTH))` STATION + DEPTH combinations have data for at least 1 full season with no major data gaps.


# Part 3: Growing Season Length

```{r, message=FALSE, echo=TRUE}

seasons_table <- dat_seasons %>% 
  group_by(STATION, SEASON, DEPTH) %>% 
  summarise(START_SEASON = min(TIMESTAMP),
            END_SEASON = max(TIMESTAMP)) %>% 
  ungroup() %>% 
  mutate(
    SEASON_DAYS = difftime(END_SEASON, START_SEASON, units = "days"),
    SEASON_DAYS = as.numeric(round(SEASON_DAYS, digits = 2)),
    SEASON_MONTHS = round(SEASON_DAYS / 30, digits = 2)
  ) %>% 
  convert_depth_to_ordered_factor() %>% 
  arrange(STATION, SEASON, DEPTH)

adcp_count_obs(seasons_table, column = "SEASON_DAYS", n_ints = 3)
adcp_count_obs(seasons_table, column = "SEASON_MONTHS", n_ints = 3)


# use above instead of identify_growing_seasons() to find start/end of each season
# can't use identify_growing_seasons() on data that is already filtered for seasons:
## for full_seasons = TRUE: returns empty dataframe because data always starts above trend_theshold and never cross superchill_threshold
## for full_seasons = FALSE, it adds additional seasons; e.g., S1 is from June 2019 to April 2020; it will add a S2 from January 1, 2020 to April 2020

```

## Visualize Growing Season Length {.tabset}

### Figure 3

```{r, echo=FALSE, message=FALSE, fig.height=8}

cols <- get_colour_palette(seasons_table)

p <- seasons_table %>%
  group_by(STATION, SEASON) %>%
  summarise(MAX_SEASON_MONTHS = max(SEASON_MONTHS)) %>%
  ggplot(aes(x = MAX_SEASON_MONTHS, y = STATION, fill = SEASON)) +
  geom_col(
    position = position_dodge2(preserve = "single", reverse = TRUE, padding = 0)
  ) +
  scale_y_discrete(name = "", limits = rev) +
  theme(panel.grid.major.y =  element_blank())

ggplotly(p)

```

### Figure 4

```{r, echo=FALSE, fig.height=12}

p <- seasons_table %>%
  ggplot(aes(x = SEASON_MONTHS, y = STATION, fill = DEPTH)) +
  geom_col(
    position = position_dodge2(preserve = "single", reverse = TRUE, padding = 0)
  ) +
  scale_fill_manual(name = "Depth (m)", values = cols) +
  scale_y_discrete(name = "", limits = rev) +
  facet_wrap(~SEASON, ncol = 1, scales = "free_y") +
  theme(panel.grid.major.y =  element_blank())

ggplotly(p)
  
```

### Table 3
```{r, echo=FALSE}

seasons_table %>%
  mutate(START_SEASON = format(START_SEASON),
         END_SEASON = format(END_SEASON)) %>% 
  arrange(STATION, SEASON, DEPTH) %>% 
  select(-SEASON_DAYS) %>% 
  datatable(rownames = FALSE, options = dt_options)

```


## Season Statistics {.tabset}

```{r, echo=FALSE}
short <- 8.5
long <- 17
```

Based on the analysis below:

   - Long season => `r long` months (red)
   - Medium season: `r short`  to < `r long` months (orange)
   - Short season: < `r short` (blue)


```{r, echo=FALSE}

seasons_table <- seasons_table %>% 
  mutate(
    SEASON_CLASS = case_when(
      SEASON_MONTHS < short ~ "Short",
      SEASON_MONTHS >= short & SEASON_MONTHS < long ~ "Medium",
      SEASON_MONTHS >= long ~ "Long"
    ),
    SEASON_CLASS = ordered(
      factor(SEASON_CLASS), levels = c("Short", "Medium", "Long")
    )
  ) %>% 
  select(STATION, SEASON, DEPTH, SEASON_MONTHS, SEASON_CLASS)

n_depths <- seasons_table %>% 
  group_by(STATION, SEASON) %>% 
  summarise(n_depths = n(), .groups = "drop" )

# add season length column
dat_final <- dat_seasons %>% 
  left_join(seasons_table, by = c("STATION", "SEASON", "DEPTH")) %>% 
  left_join(n_depths, by = c("STATION", "SEASON")) %>% 
  filter(n_depths > 1) %>% 
  select(-n_depths)

```

### Figure 5
```{r, echo=FALSE}
p <- ggplot(seasons_table, aes(x = SEASON_MONTHS, fill = DEPTH)) +
  geom_histogram(binwidth = 0.5) +
  scale_fill_manual(name = "Depth (m)", values = cols) +
  scale_x_continuous("Season Length (months)") +
  labs(title = "Histogram of season length (all STATION, SEASON, DEPTHs)")

ggplotly(p)

```

### Figure 6
```{r, echo=FALSE}
p <- ggplot(seasons_table, aes(y = SEASON_MONTHS)) +
  geom_boxplot() +
  scale_y_continuous("Season Length (months)") +
  labs(title = "Boxplot of season length (all STATION, SEASON, DEPTHS)") +
  theme(axis.text.x = element_blank())
  
ggplotly(p)

```

### Figure 7
```{r, echo=FALSE}
p <- seasons_table %>% 
  filter(SEASON_MONTHS < 17) %>% 
  ggplot(aes(x = SEASON_MONTHS, fill = DEPTH)) +
  geom_histogram(binwidth = 0.5) +
  scale_fill_manual(name = "Depth (m)", values = cols) +
  scale_x_continuous("Season Length (months)") +
  labs(title = "Season length < 17 months")

ggplotly(p)
```

### Figure 8
```{r, echo=FALSE}
p <- seasons_table %>% 
  filter(SEASON_MONTHS < 17) %>% 
  ggplot(aes(y = SEASON_MONTHS)) +
  geom_boxplot() +
  scale_y_continuous("Season Length (months)") +
  labs(title = "Season length < 17 months") +
  theme(axis.text.x = element_blank())
  
ggplotly(p)

```

### Figure 9
```{r, echo=FALSE}
p <- seasons_table %>% 
  ggplot(aes(x = SEASON_CLASS, y = SEASON_MONTHS)) +
  geom_boxplot() +
  scale_x_discrete("") +
  scale_y_continuous("Season Length (months)") 

ggplotly(p)

```


# Part 4: Visualize Data Series 

```{r, echo=FALSE}

rad <- 8
alpha = 0.5

dat_map <- dat_final %>% 
  select(STATION, SEASON, SEASON_CLASS, LATITUDE, LONGITUDE) %>% 
  distinct(STATION, SEASON, .keep_all = TRUE) 

leaflet(dat_map) %>%
  addProviderTiles(providers$CartoDB.Positron) %>%
  addCircleMarkers(
    data = filter(dat_map, SEASON_CLASS == "Short"),
    ~LONGITUDE, ~LATITUDE, label = ~STATION,
    fillColor = "blue",
    stroke = FALSE,
    fillOpacity = alpha,
    radius = rad
  ) %>% 
  addCircleMarkers(
    data = filter(dat_map, SEASON_CLASS == "Medium"),
    ~LONGITUDE, ~LATITUDE, label = ~STATION,
    fillColor = "orange",
    stroke = FALSE,
    fillOpacity = alpha,
     radius = rad
  ) %>% 
  addCircleMarkers(
    data = filter(dat_map, SEASON_CLASS == "Long"),
    ~LONGITUDE, ~LATITUDE, label = ~STATION,
    fillColor = "red",
    stroke = FALSE,
    fillOpacity = alpha,
     radius = rad
  )
  
```

NOTE: Figures only show data for FULL seasons.

```{r, echo=FALSE, message=FALSE, results='asis'}

season_cat  <- c("Long", "Medium", "Short")

for(j in seq_along(season_cat)){
  
  cat('\n##', season_cat[j], "Seasons {.tabset} \n")
  
  stations_j <- dat_final %>%
    filter(SEASON_CLASS == season_cat[j]) %>% 
    select(STATION) %>% 
    distinct()
  stations_j <- sort(stations_j$STATION)
  
  for(i in seq_along(stations_j)){
    
    cat('\n###', stations_j[i], '\n')
    
    cat('\n')
    
    p <- dat_final %>% 
      filter(SEASON_CLASS == season_cat[j], 
             STATION == stations_j[i]) %>% 
      plot_temperature_at_depth(facet_var = "SEASON",
                                date_labels_format = "%Y-%b") 
    
    print(p)
    
    cat('\n')
    
  }
}
```
