==> R/3_figures_interactive.Rmd <==
---
title: 'Figures'
author: "Danielle Dempsey"
date: "`r Sys.Date()`"
output:
  html_document: default
editor_options:
  chunk_output_type: console
---
  
```{r, echo = FALSE, message=FALSE}

knitr::opts_chunk$set(echo = FALSE, fig.width=9, message=FALSE)

library(dplyr)         # data manipulation 
library(DT)            # interactive tables
library(ggplot2)       # figures
library(ggspatial)     # for basemap
library(glue)          # used in map figure
library(here)          # relative files path 
library(lubridate)     # dates
library(RColorBrewer)  # for TGC model figure
library(sf)            # static map 
library(strings)       # convert_depth_to_ordered_factor() function
library(tgc)           # season functions
library(plotly)        # interactive figures
library(viridis)       # colour palette

dt_options <- list(
      dom = 'ft',
      paging = FALSE,
      searching = TRUE,
      scrollY = "500px",
      scrollX = "500px",
      pageLength = 500,
      columnDefs = list(list(className = 'dt-center', targets = "_all"))
)

theme_set(theme_light())

```

```{r}

results <- readRDS(here("results/model_results.rds"))

dat_raw <- results$dat_raw
dat_seasons <- results$dat_seasons
dat_filt <- results$dat_filt
gap_table <- results$gap_table
dd <- results$dd
tgc_table <- results$tgc_table

# plot params
i <- 1
n_days <- 14
text_size <- 8
plot_theme <- theme(axis.text = element_text(size = text_size))

colour_pal <- viridis(5, direction = -1)

```

# Methods

```{r, echo=FALSE, out.width="15%", out.height="15%"}
knitr::include_graphics("Fig1.png")
```

Figure `r i`: Conceptual diagram of a sensor string (not to scale). Each string is anchored to the and suspended by a sub-surface buoy, with autonomous sensors attached at various depths.

```{r}
i <- i + 1

dd_input <- seq(500, 5000, 1)

dd_example <- data.frame(n_degree_days = dd_input)

w0 <- TGC_calculate_initial_weight(
  dd_example, final_weight = 5.5, tgc = c(0.25, 0.30, 0.35)
)

# w0 <- w0 %>% 
#   mutate(dw = 3 * (1.83 - TGC/1000*n_degree_days)^2)

p1 <- ggplot(w0, aes(n_degree_days, TGC_INITIAL_WEIGHT, col = factor(TGC))) +
  geom_line(size = 1.5) +
  scale_x_continuous("Number of Degree Days") +
  scale_y_continuous("Initial Weight (kg)") +
  scale_colour_manual("TGC Value", values = brewer.pal(3, "Dark2")) +
  plot_theme +
  theme(
    legend.position = c(0.85, 0.8),
    legend.background = element_rect(fill = "white", colour = "darkgrey")
  ) 

ggplotly(p1)

```
Figure `r i`: Initial weight required for salmon to grow to market size (5.5 kg) for three TGC values.
(Better way to phrase this??)

# Results

```{r, echo=FALSE}

i <- i + 1

dat_map <- dat_raw %>%
  distinct(STATION, .keep_all = TRUE) %>%
  mutate(
    SEASON = case_when(
      STATION == "Flat Island" ~ "Long Season",
      STATION == "Beaver Point" ~ "Medium Season",
      STATION == "Madeline Point" ~ "Short Season"
    ),
    LABEL = glue("{STATION} \n ({SEASON})")) %>%
  st_as_sf(coords = c("LONGITUDE", "LATITUDE"), crs = 4326, agr = "constant")

map_box <- data.frame(Long = c(-66, -59.9), Lat = c(43.5, 47)) %>%
  st_as_sf(coords = c("Long", "Lat"), crs = 4326)

ggplot() +
  geom_sf(data = map_box) +
  annotation_map_tile(type = "cartolight", zoomin = -1) +
  geom_sf(data = dat_map) +
  geom_sf_label(
    data = dat_map, aes(label = LABEL),
    label.size = NA, fill = NA, nudge_y = -0.25, size = 2.7
  ) +
  scale_x_continuous(breaks = c(seq(-60, -66, -2))) +
  scale_y_continuous(breaks = c(seq(44, 47, 1))) +
  annotation_scale(location = "br") +
  annotation_north_arrow(
    location = "tl", which_north = "true",
    height = unit(1, "cm"),
    width = unit(1, "cm")
  ) +
  theme(axis.title = element_blank(),
        text = element_text(size = 10))
```
Figure `r i`: Stations selected to represent a short, medium, and long growing season.


## Growing Seasons {.tabset}

`r i <- i + 1`

### Figure `r i`A

```{r, echo=FALSE, message=FALSE}

short_filt <- dat_filt %>% filter(STATION == "Madeline Point")

min_timestamp <- min(short_filt$TIMESTAMP)
max_timestamp <- max(short_filt$TIMESTAMP)

short_raw <- dat_raw %>% 
  filter(
    STATION == "Madeline Point",
    TIMESTAMP >= min_timestamp & TIMESTAMP <= max_timestamp
  )

p2 <- plot_filtered_data(
  short_raw, short_filt,
  colour_palette = colour_pal,
  date_breaks_major = "1 month",
  date_labels_format = "%Y-%b",
  plotly_friendly = FALSE
) +
  plot_theme

p2

```

### Figure `r i`B

```{r, echo=FALSE, message=FALSE}

med_filt <- dat_filt %>% filter(STATION == "Beaver Point")

min_timestamp <- min(med_filt$TIMESTAMP) 
max_timestamp <- max(med_filt$TIMESTAMP)

med_raw <- dat_raw %>% 
  filter(
    STATION == "Beaver Point",
    TIMESTAMP >= min_timestamp & TIMESTAMP <= max_timestamp
  )

p2 <- plot_filtered_data(
  med_raw, med_filt,
  colour_palette = colour_pal,
  date_breaks_major = "1 month",
  date_labels_format = "%Y-%b",
  plotly_friendly = FALSE) +
  plot_theme

p2

```

### Figure `r i`C

```{r, echo=FALSE, message=FALSE}

long_filt <- dat_filt %>% filter(STATION == "Flat Island")

min_timestamp <- min(long_filt$TIMESTAMP) 
max_timestamp <- max(long_filt$TIMESTAMP)

long_raw <- dat_raw %>% 
  filter(
    STATION == "Flat Island",
    TIMESTAMP >= min_timestamp & TIMESTAMP <= max_timestamp
  )

p2 <- plot_filtered_data(
  long_raw, long_filt,
  colour_palette = colour_pal,
  date_breaks_major = "2 month",
  date_labels_format = "%Y-%b",
  plotly_friendly = FALSE) +
  plot_theme

p2

```

Figure `r i`: Temperature at each example station and depth. The blue shaded region indicates the superchill threshold, the red shaded area indicates the heat stress threshold, and the grey dashed line indicates the 4-degree threshold. Greyed observations were filtered out of the analysis.


## Degree days & TGC

```{r}
i <- i+1

p <- ggplot(dd, aes(STATION, n_degree_days, fill = DEPTH,
                    text = paste(
                      "Depth: ", DEPTH, "m", "\n",
                      "Average Temp: ", AVG_TEMPERATURE, "\n",
                      "Growing Days: ", n_growing_days, "\n",
                      "Degree Days: ", n_degree_days))
) +
  geom_point(pch = 21, size = 4, alpha = 0.75,
             position = position_jitter(width = 0, height = 0, seed = 10)) +
  scale_fill_manual("Depth (m)", values = colour_pal) +
  scale_x_discrete(name = "") +
  scale_y_continuous(name = "Number of Degree Days") +
  theme_light()

ggplotly(p, tooltip = "text") %>%
  config(
    modeBarButtonsToRemove = list(
      "toggleSpikelines", "autoScale2d", "select2d", "lasso2d",
      "hoverCompareCartesian", "zoom2d"
    ),
    displaylogo = FALSE,
    toImageButtonOptions = list(
      format = "png",
      filename = "degree_days",
      width = 1100,
      height = 1100
    )) %>%
  layout(
    margin = list(b = 90, l = 90) # to fully display the x and y axis labels
  )

```
Figure `r i`: Number of growing degree days at each example station and depth.


Table 1: Growing season and degree day results for each example station and depth. (Note: Partial STOCKED_DAYS caused by the times of END_SEASON and START_SEASON; not shown).
```{r echo=FALSE}

dd %>%
  mutate(
    START_SEASON = format(as_date(START_SEASON)),
    END_SEASON = format(as_date(END_SEASON))
  ) %>% 
  arrange(STATION) %>% 
  datatable(
    rownames = FALSE, options = dt_options
  )

```


```{r}
i <- i +1

p <-  ggplot(
  tgc_table, aes(x = factor(TGC), y = TGC_INITIAL_WEIGHT, fill = DEPTH,
             text = paste(
                      "Depth: ", DEPTH, "m", "\n",
                      "Average Temp: ", AVG_TEMPERATURE, "\n",
                      "Growing Days: ", n_growing_days, "\n",
                      "Degree Days: ", n_degree_days, "\n",
                      "Initial Weight: ", TGC_INITIAL_WEIGHT ))) +
  geom_point(
    pch = 21, size = 3, alpha = 0.75,
    position = position_jitter(width = 0, height = 0, seed = 10)
  ) +
  scale_fill_manual("Depth (m)", values = colour_pal, drop = FALSE) +
  scale_x_discrete("TGC Value") +
  scale_y_continuous("Initial Weight (kg)") +
  facet_wrap(~STATION)


ggplotly(p, tooltip = "text") %>%
  config(
    modeBarButtonsToRemove = list(
      "toggleSpikelines", "autoScale2d", "select2d", "lasso2d",
      "hoverCompareCartesian", "zoom2d"
    ),
    displaylogo = FALSE,
    toImageButtonOptions = list(
      format = "png",
      filename = "degree_days",
      width = 1100,
      height = 1100
    )) %>%
  layout(
    margin = list(b = 90, l = 90) # to fully display the x and y axis labels
  )

```

Figure `r i`: TGC model results.
